fix: re-add same-incident field to incident overview
