+++ keep-ui/features/same-incidents-in-the-past/index.ts
@@ -1,1 +1,1 @@
-export { SameIncidentsOverview } from "./ui/same-incidents-overview";
+export { SameIncidentsOverview } from "./ui/following-incidents";
