+++ keep-ui/features/same-incidents-in-the-past/ui/index.ts
@@ -1,0 +1,2 @@
+export { SameIncidentField } from "./same-incident-field";
+export { FollowingIncidents } from "./following-incidents";
