+++ keep-ui/shared/ui/FieldHeader.tsx
@@ -1,0 +1,3 @@
+export const FieldHeader = ({ children }: { children: React.ReactNode }) => (
+  <h3 className="text-sm text-gray-500 font-semibold">{children}</h3>
+);
