+++ keep-ui/shared/ui/index.ts
@@ -1,2 +1,4 @@
 export { TablePagination } from "./TablePagination";
 export { TabLinkNavigation, TabNavigationLink } from "./TabLinkNavigation";
+export { DateTimeField } from "./DateTimeField";
+export { FieldHeader } from "./FieldHeader";
